Do a bad job of rendering only the selected options using Options class
The `onClick` handler that I've passed to each button only *mutates* the
button inplace.
This has the effect of mutating the object but not triggering a rerender from
React, and so in this commit I've first, mutated the object, and then terribly
forced a rerender of the SAME (now-mutated) array to React, which causes it to
draw the updated state.
I'm sure there's a better way to do this, even though it techically works.