Scene 2, Take 1 of fuzzy category tags
Feel there's a better way to do it, but might come back and fix it in the
future.
Unlike previous attempts, this one foregoes use of FuzzySearchBar for the
time being, and rather just has it's own text field and search function.
The reason I couldn't use FSB is that FSB's current design requires it to
be able to set the state of the search objects after search on it, but I
specifically didn't want to introduce more complex state maintenance logic
after my last PR to remove that from CategoryTags.
In the future, we should revisit FSB to make it better.